+++ src/webgpu/api/validation/render_pipeline/inter_stage.spec.ts
@@ -1,0 +1,10 @@
+export const description = `
+TODO: interface matching between vertex and fragment shader validation for createRenderPipeline:
+    - superset, subset, etc.
+`;
+
+import { makeTestGroup } from '../../../../common/framework/test_group.js';
+
+import { CreateRenderPipelineValidationTest } from './common.js';
+
+export const g = makeTestGroup(CreateRenderPipelineValidationTest);
